part 4 - completed exercise 4.22

+++ tests/blog_api.test.js
@@ -3,6 +3,8 @@ const supertest = require('supertest');
 const app = require('../app');
 const blogAPITestHelper = require('./blog_api_test_helpers');
 const Blog = require('../models/blog');
+const User = require('../models/user');
+
 
 const api = supertest(app);
 
@@ -17,62 +19,69 @@ test('there exists an id property', async () => {
 	expect(blogs[0].id).toBeDefined();	
 });
 
-test('a valid blog can be added', async () => {
-	const newBlog = {
-        title: "Biden vs Trump - Continued",
-        author: "Nate Silver",
-        url: "https://fivethirtyeight.com",
-        likes: 200
-	};
+const getJwtToken = async () => {
+	const testUserCreds = {username: 'testUser1', password: 'password'};
+	const userLogin = await api.post('/api/login').send(testUserCreds).expect(200);
+	return userLogin.body.token;
+}
+
+describe('creating a blog', () => {
+	test('a valid blog can be added', async () => {		
+
+		const newBlog = {
+			title: "Biden vs Trump - Continued",
+			author: "Nate Silver",
+			url: "https://fivethirtyeight.com",
+			likes: 200
+		};
+		
+		await api.post('/api/blogs').send(newBlog).set('authorization', `Bearer ${await getJwtToken()}`).expect(201);
 	
-	await api.post('/api/blogs').send(newBlog).expect(201);
-
-	const response = await api.get('/api/blogs');
-
-	console.log('blogs response', response.body);
-	const contents = response.body.map(r => r.title);
-
-	expect(response.body).toHaveLength(blogAPITestHelper.initialBlogs.length + 1);
-
-	expect(contents).toContain(
-		'Biden vs Trump - Continued'
-	);
-});
-
-test('a blog without likes key', async () => {
-	const newBlog = {
-        title: "Biden vs Trump - Continued",
-        author: "Nate Silver",
-        url: "https://fivethirtyeight.com"
-	};
+		const response = await api.get('/api/blogs');
 	
-	await api.post('/api/blogs').send(newBlog).expect(201);
-
-	const response = await api.get('/api/blogs');
-
-	console.log('blogs response', response.body);
-	const blog = response.body.find(r => r.title === newBlog.title);
-
-	expect(response.body).toHaveLength(blogAPITestHelper.initialBlogs.length + 1);
-
-	expect(blog.likes).toBe(0);
-});
-
-test('post api should return 400 status when title/url missing', async () => {
-	const newBlog = {
-        author: "Nate Silver",
-        url: "https://fivethirtyeight.com"
-	};
+		const contents = response.body.map(r => r.title);
 	
-	await api.post('/api/blogs').send(newBlog).expect(400);
+		expect(response.body).toHaveLength(blogAPITestHelper.initialBlogs.length + 1);
+	
+		expect(contents).toContain(
+			'Biden vs Trump - Continued'
+		);
+	});
+	
+	test('a blog without likes key', async () => {
+		const newBlog = {
+			title: "Biden vs Trump - Continued",
+			author: "Nate Silver",
+			url: "https://fivethirtyeight.com"
+		};
+		
+		await api.post('/api/blogs').send(newBlog).set('authorization', `Bearer ${await getJwtToken()}`).expect(201);
+	
+		const response = await api.get('/api/blogs');
+	
+		const blog = response.body.find(r => r.title === newBlog.title);
+	
+		expect(response.body).toHaveLength(blogAPITestHelper.initialBlogs.length + 1);
+	
+		expect(blog.likes).toBe(0);
+	});
+	
+	test('post api should return 400 status when title/url missing', async () => {
+		const newBlog = {
+			author: "Nate Silver",
+			url: "https://fivethirtyeight.com"
+		};
+		
+		await api.post('/api/blogs').send(newBlog).set('authorization', `Bearer ${await getJwtToken()}`).expect(400);
+	});
 });
 
 describe('deletion of a blog', () => {
 
 	test('succeeds with status code 204 if id is valid', async () => {
 		const blogsResponse = await api.get('/api/blogs');
-		const blogToDelete = blogsResponse.body[blogsResponse.body.length - 1]; //last blog		
-		await api.delete(`/api/blogs/${blogToDelete.id}`).expect(204);
+		const blogToDelete = blogsResponse.body[blogsResponse.body.length - 1]; //last blog
+		await api.delete(`/api/blogs/${blogToDelete.id}`).set('authorization', `Bearer ${await getJwtToken()}`).expect(204);
 
 		const newBlogsResponse = await api.get('/api/blogs');
 		const titles = newBlogsResponse.body.map(r => r.title);
@@ -83,7 +92,7 @@ describe('deletion of a blog', () => {
 
 	test('fails with status code 404 if id is invalid', async () => {
 		const invalidBlogId = "5fa64876e60b7b4be14a2d37";		
-		await api.delete(`/api/blogs/${invalidBlogId}`).expect(404);
+		await api.delete(`/api/blogs/${invalidBlogId}`).set('authorization', `Bearer ${await getJwtToken()}`).expect(404);
 	});
 
 })
@@ -95,6 +104,7 @@ describe('updating a blog', () => {
 		const blogToUpdate = blogsResponse.body[0]; //first blog
 		const newLikesCount = 0;
 		blogToUpdate.likes = newLikesCount;
+		delete blogToUpdate['user'];
 		
 		await api.put(`/api/blogs/${blogToUpdate.id}`).send(blogToUpdate).expect(200);
 
@@ -116,14 +126,27 @@ describe('updating a blog', () => {
 
 beforeEach(async () => {
 	await Blog.deleteMany({});//empty the collection
-	console.log('cleared')
+	await User.deleteMany({});
+	
+	const userPayload = {
+		username: 'testUser1',
+		password: 'password',
+		name: 'Test User 1'
+	};
 
-	console.log('inserting new blogs');
-	const newBlogs = blogAPITestHelper.initialBlogs.map(blog => new Blog(blog))
+	await api.post('/api/users').send(userPayload);
+
+	const user = await User.findOne({username: userPayload.username});
+
+	const newBlogs = blogAPITestHelper.initialBlogs.map(blog => new Blog({
+		...blog,
+		user: user._id
+	}));
+
   	const promiseArray = newBlogs.map(blog => blog.save());
-  	await Promise.all(promiseArray);
-
-	console.log('done')
+  	const allBlogs = await Promise.all(promiseArray);
+	user.blogs = allBlogs.map(blog => blog._id);
+	await user.save();
 })
 
 afterAll(() => {
